Move ExprPtr related things to expr_ptr.cpp
